Pass size_t to span functions: services/
In preparation for upgrading various span args from size_t to
StrictNumeric<size_t>, this fixes type mismatches that would cause
compile errors with that change.

Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6020233
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: David Schinazi <dschinazi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1382906}

+++ services/network/public/cpp/net_adapters.h
@@ -82,7 +82,7 @@ class COMPONENT_EXPORT(NETWORK_CPP) NetToMojoIOBuffer
   // will be offset by that many bytes.
   explicit NetToMojoIOBuffer(
       scoped_refptr<NetToMojoPendingBuffer> pending_buffer,
-      int offset = 0);
+      size_t offset = 0);
 
   NetToMojoIOBuffer(const NetToMojoIOBuffer&) = delete;
   NetToMojoIOBuffer& operator=(const NetToMojoIOBuffer&) = delete;
